+++ task09-reactjs/task09/src/components/Header.js
@@ -2,14 +2,24 @@
  * All Rights Reserved
  * This software is property information of AKUREY S.A.
  * Contact us at contact@akurey.com
- * @summary short description for the file
+ * @summary js for the header component
  */
 
- import '../scss/Header.scss';
-
- function NavbarList() {
-    return (
-            <div className="collapse navbar-collapse" id="navbarTogglerDemo01">
+import '../scss/Body.scss';
+function Header() {
+  return (
+    
+      <header>
+        <nav className={`navbar fixed-top navbar-expand-xl navbar-light bg-white py-4 justify-content-center`}>
+          <button className='navbar-toggler burger-menu' type='button' data-toggle="collapse" data-target="#navbarTogglerDemo01"
+          aria-controls="navbarTogglerDemo01" aria-expanded="false" aria-label="Toggle navigation">
+            <span id="burger-icon" className="icon-align-left"></span>
+            <span id="burger-icon" className="icon-times"></span>
+          </button>
+          <div className="logo">
+            <img className="one-piece-logo" src={require("../images/one-piece-logo.png")} alt="One Piece Logo" />
+          </div>
+          <div className="collapse navbar-collapse" id="navbarTogglerDemo01">
               <ul className="navbar-nav">
                 <li className="nav-item">
                   <a className="nav-link active" href="/">Home<span class="sr-only">(current)</span></a>
@@ -35,8 +45,10 @@
                 </div>
               </div>
             </div>
-    );
-  }
-  
-  export default NavbarList;
-  
+        </nav>
+      </header>
+    
+  );
+}
+
+export default Header;
