+++ task09-reactjs/task09/src/scss/App.scss
@@ -1,6 +1,9 @@
-.App {
-  text-align: center;
-}
+/**
+* All Rights Reserved
+* This software is property information of AKUREY S.A.
+* Contact us at contact@akurey.com
+* @summary scss with styling/mixins for App.js
+*/
 
 .App-logo {
   height: 40vmin;
