+++ task09-reactjs/task09/src/scss/_gridflex.scss
@@ -1,3 +1,9 @@
+/**
+* All Rights Reserved
+* This software is property information of AKUREY S.A.
+* Contact us at contact@akurey.com
+* @summary scss with styling/mixins for grid an flex
+*/
 @mixin grid-format( $columns, $gap:0) {
     display: grid;
     grid-template-columns: $columns;
