+++ task09-reactjs/task09/src/scss/index.scss
@@ -1,3 +1,10 @@
+/**
+* All Rights Reserved
+* This software is property information of AKUREY S.A.
+* Contact us at contact@akurey.com
+* @summary scss with styling/mixins for index.js
+*/
+
 body {
   padding-top: 88px;
 }
